requirements.txt
pandas
python-docx
pdfplumber
psycopg2-binary
openai
crewai
langchain
python-dotenv
watchdog
